requirements.txt
PyQt6>=6.5,<7.0
icalendar>=5.0,<6.0
python-dateutil>=2.8,<3.0
